==> scripts/cardinalR/five_clusts/02_gen_embeddings.R <==
library(dplyr)
library(tibble)
library(readr)
library(conflicted)

library(Rtsne)
library(uwot)
library(phateR)
library(reticulate)

set.seed(20240110)

use_python("~/miniforge3/envs/pcamp_env/bin/python")
use_condaenv("pcamp_env")


data <- read_rds(here::here("data/five_clusts/five_clusts.rds"))
data_n <- data
data <- data |>
  dplyr::select(-cluster)

## tSNE (default)
perplexity <- 30

tSNE_fit <- data |>
  dplyr::select(where(is.numeric)) |>
  Rtsne::Rtsne(perplexity = perplexity,
               pca = FALSE)

tSNE_data <- tSNE_fit$Y |>
  tibble::as_tibble(.name_repair = "unique")
names(tSNE_data) <- c("tSNE1", "tSNE2")

tSNE_data <- tSNE_data |>
  dplyr::mutate(cluster = data_n$cluster)

write_rds(tSNE_data, file = paste0("data/five_clusts/five_clusts_tsne_perplexity_", perplexity, ".rds"))

## UMAP

n_neighbors <- 15
min_dist <- 0.1

UMAP_model <- umap(data,
                  n_neighbors = n_neighbors,
                  min_dist = min_dist,
                  n_components =  2,
                  init ="spca")

UMAP_data <- UMAP_model |>
  as_tibble(.name_repair = "unique")

names(UMAP_data) <- c("UMAP1", "UMAP2")

UMAP_data <- UMAP_data |>
  dplyr::mutate(cluster = data_n$cluster)

## Run only once
write_rds(UMAP_data, file = paste0("data/five_clusts/five_clusts_umap_n-neigbors_", n_neighbors, "_min-dist_", min_dist, ".rds"))

## PHATE
knn <- 5

PHATE_data <- phate(data, knn = knn)
PHATE_data <- as_tibble(PHATE_data$embedding)

names(PHATE_data) <- c("PHATE1", "PHATE2")

PHATE_data <- PHATE_data |>
  dplyr::mutate(cluster = data_n$cluster)

write_rds(PHATE_data, file = paste0("data/five_clusts/five_clusts_phate_knn_", knn, ".rds"))


## TriMAP

trimap <- reticulate::import("trimap")

data_vector <- unlist(data)
# Convert the vector into a matrix
data_matrix <- matrix(data_vector, ncol = NCOL(data))

n_inliers <- as.integer(12)
n_outliers <- as.integer(4)
n_random <- as.integer(3)

# Initialize PaCMAP instance
reducer <- trimap$TRIMAP(n_dims = as.integer(2),
                         n_inliers = n_inliers,
                         n_outliers = n_outliers,
                         n_random = n_random)

# Perform dimensionality Reduction
TriMAP_data <- reducer$fit_transform(data_matrix) |>
  as_tibble()

names(TriMAP_data) <- c("TriMAP1", "TriMAP2")

TriMAP_data <- TriMAP_data |>
  dplyr::mutate(cluster = data_n$cluster)

write_rds(TriMAP_data, file = paste0("data/five_clusts/five_clusts_trimap_n-inliers_", n_inliers, "_n-outliers_", n_outliers, "_n-random_", n_random, ".rds"))

## PaCMAP

pacmap <- reticulate::import("pacmap")

data_vector <- unlist(data)
# Convert the vector into a matrix
data_matrix <- matrix(data_vector, ncol = NCOL(data))

n_neighbors <- as.integer(10)
MN_ratio <- 0.5
FP_ratio <- as.integer(2)
init <- "random"

# Initialize PaCMAP instance
reducer <- pacmap$PaCMAP(n_components = as.integer(2),
                         n_neighbors = n_neighbors,
                         MN_ratio = MN_ratio,
                         FP_ratio = FP_ratio)


# Perform dimensionality Reduction
PacMAP_data <- reducer$fit_transform(data_matrix, init = init) |>
  as_tibble()

names(PacMAP_data) <- c("PaCMAP1", "PaCMAP2")

PacMAP_data <- PacMAP_data |>
  dplyr::mutate(cluster = data_n$cluster)

write_rds(PacMAP_data, file = paste0("data/five_clusts/five_clusts_pacmap_n-neighbors_", n_neighbors,"_init_", init, "_MN-ratio_", MN_ratio, "_FP-ratio_", FP_ratio, ".rds"))

## PCA

calculate_pca <- function(feature_dataset){
  pcaY_cal <- prcomp(feature_dataset, center = TRUE, scale = FALSE)
  PCAresults <- data.frame(pcaY_cal$x[, 1:2])
  summary_pca <- summary(pcaY_cal)
  var_explained_df <- data.frame(PC= paste0("PC",1:2),
                                 var_explained=(pcaY_cal$sdev[1:2])^2/sum((pcaY_cal$sdev[1:2])^2))
  return(list(prcomp_out = pcaY_cal,pca_components = PCAresults, summary = summary_pca, var_explained_pca  = var_explained_df))
}

pca_ref_calc <- calculate_pca(data)
data_pca <- pca_ref_calc$pca_components |>
  stats::setNames(c("pca1", "pca2"))

data_pca <- data_pca |>
  dplyr::mutate(cluster = data_n$cluster)

write_rds(data_pca, file = paste0("data/five_clusts/five_clusts_pca.rds"))
